abstract common undolog manager method
